feat: add opening order info as a modal window
