+++ src/components/modal/modal.module.css
@@ -18,6 +18,7 @@
   opacity: 0;
   visibility: hidden;
   transition: opacity .2s, visibility .2s linear;
+  z-index: 999;
 }
 
 .show {
