+++ src/pages/feed/feed.module.css
@@ -41,6 +41,7 @@
   display: flex;
   flex-direction: column;
   gap: 24px;
+  cursor: pointer;
 }
 
 .orderDetailsHeader {
