+++ src/pages/orderInfo/orderInfo.module.css
@@ -5,82 +5,3 @@
   justify-content: center;
   height: calc(100vh - 88px);
 }
-
-.orderInfoSection {
-  width: 100%;
-  max-width: 640px;
-}
-
-.orderNumber {
-  text-align: center;
-}
-
-.ingredientsArea {
-  max-height: 312px;
-  overflow: hidden;
-  overflow-y: auto;
-}
-
-.ingredientsList {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-}
-
-.ingredient {
-  width: 100%;
-  min-height: 64px;
-  display: flex;
-  flex-direction: row;
-  gap: 16px;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.ingredientName {
-  width: 100%;
-}
-
-.ingredientImageContainer {
-  width: 52px;
-  height: 52px;
-  padding: 4px;
-  border-radius: 100%;
-  border: 2px solid transparent;
-  background:
-    linear-gradient(#1C1C21, #1C1C21) padding-box,
-    linear-gradient(45deg,#801AB2, #4C4CFF) border-box;
-  object-fit: contain;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  position: relative;
-}
-
-.ingredientImage {
-  height: 100%;
-}
-
-.ingredientPrice {
-  width: 260px;
-  display: flex;
-  flex-direction: row;
-  gap: 8px;
-  justify-content: flex-end;
-}
-
-.orderInfoFooter {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.orderTotal {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-}
